rename path to lens
