+++ src/config/reader_context.rs
@@ -2,7 +2,7 @@ use std::borrow::Cow;
 use std::collections::BTreeMap;
 use std::sync::Arc;
 
-use crate::path::LensPath;
+use crate::lens::LensPath;
 use crate::EnvIO;
 
 pub struct ConfigReaderContext<'a> {
