+++ src/lib.rs
@@ -20,9 +20,9 @@ pub mod helpers;
 pub mod http;
 pub mod json;
 pub mod lambda;
+pub mod lens;
 pub mod merge_right;
 pub mod mustache;
-pub mod path;
 pub mod print_schema;
 mod rest;
 pub mod runtime;
